requirements.txt
fastmcp>=3.0.0
httpx>=0.27.0
Pillow>=10.0.0
